Move Quantity to sub-directory + breakout the subcomponent
