+++ ndsl/quantity/__init__.py
@@ -1,0 +1,9 @@
+from ndsl.quantity.quantity import Quantity
+from ndsl.quantity.metadata import QuantityMetadata, QuantityHaloSpec
+
+
+__all__ = [
+    "Quantity",
+    "QuantityMetadata",
+    "QuantityHaloSpec",
+]
